Move to it's own class.
